Need to switch to readlink
It's odd, but you can have the following directory structure

	.git/hooks -> ../util/git/hooks/
	bin/
	util/bin/
	util/git/hooks/

In util/git/hooks/post-checkout you put in the following things and then
run the command in the top of the repository.

    . "$(dirname "$0")/../../bin/TESTME"
    cd "$(dirname "$0")/../../bin/"

It gets resolved like you did this

    . "util/bin/TESTME"
	cd "bin/"

It is clear that the first one, where the file is sourced, follows the
symbolic link into util/git/hooks and the second resolves the path by
skipping directory traversal.  So I use readlink to force the path to be
fully resolved and that works better for the setup script.

Added tests and fixed a minor bug that affects this issue as well.