Adding php-cs-fixer the rest of the way
Renamed the file to match current conventions.  Underscores are used in
the options and bash variables, so they are avoided for names of the
pretty printers and lint checkers.

Added the pretty printer to the documentation.

Set up options and included the pretty printer in the list of defaults.

Must disable "PSR-0" fixes because that messes with the name of the
temporary file during pretty printing and doesn't change the original
file.  git-started is not set up to handle this sort of situation, and
I feel it's outside the scope of a typical "pretty printer" program to
rename files.